+++ src/main/java/com/telepathicgrunt/commandstructures/CommandStructuresMain.java
@@ -20,5 +20,6 @@ public class CommandStructuresMain {
 
     private void registerCommand(RegisterCommandsEvent event) {
         StructureSpawnCommand.dataGenCommand(event.getDispatcher());
+        SpawnPiecesCommand.dataGenCommand(event.getDispatcher());
     }
 }